Implement imageRootDirMounted and aufsDiffPaths
imageRootDir now works for aufs, but shows only the last diff in the image.
To have fully compatible behavior, imageRootDirMounted should be used, so that
the full image rootfs gets mounted at the returned path. For other drivers,
behavior will be the same as imageRootDir (only it has to be used with Promise.using).

aufsDiffPaths allows getting the paths for each layer diff in an aufs-stored image.