Fix mounting aufs when there's several layers and use random path for mount.
mount limits the mount options to page size (usually 4096 bytes), so we do what docker does
to mount when layers don't fit in that: we mount with as much as we can fit and then
remount appending each remaining layer. We stick to 4096 instead of actually finding the page size
cause it's assured to work (4096 is the minimum in the arches we support), reduces complexity and
shouldn't generate much overhead (we don't care about a few extra remounts).

Based on https://github.com/docker/docker/blob/b937aa8e6968d805527d163e6f477d496ceb88d7/daemon/graphdriver/aufs/aufs.go#L595

We also introduce 8 random characters to the temporary path we use for the aufs mount, so that
several calls to this function can work independently.

Change-Type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>