Added support for inspecting and mounting overlay2 fs
The overlay2 driver adds support for more than one “lower” layer (quite similar to aufs) and changed how an OverlayFS is represented on disk. This commit adds support for inspecting the “diff” layer and mounting an fs created with overlay2.

This deprecates `aufsDiffPaths`, which now also applies to overlay2, introducing a new method `diffPaths` that can work with both aufs and overlay2.

Change-Type: minor