Use dockerode's promise functionality (docker will not have Async-suffixed methods)
This module used to promisify and mutate the Docker class from dockerode, which is bad.
It also mutated images and containers by promisifying their prototype.
Now we extend that class and use dockerode's promise support instead of promisifying - but we pass bluebird
as the Promise for dockerode to use.

We also add mocha (with mochainon) and some basic tests that only cover testing that this returns a valid promise-aware
docker object.
No tests for the docker-toolbelt specific functions were added yet (should be added soon, hopefully...).

Change-Type: major
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>