Allow getRegistryAndName, compileRegistryAndName, and normaliseImageName to work with images with digest
We add an additional "digest" field to the object returned by getRegistryAndName, which is populated  with the digest
when the image has digest instead of tag. Otherwise behavior us the same as before.
compileRegistryAndName now accepts a digest field too.

normaliseImageName worked with digests but it was just lucky, as getRegistryAndName would match the "@sha256" as part of the repo.
This change makes the support intentional, and matching the image name in the correct way.

Change-Type: minor
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>